[Fix]Fixed minor bugs.

+++ cms/common/site_include/plugin/util_multi_language/func/fn.php
@@ -3,37 +3,39 @@ function multi_language_execute_common_update_process(int $id, string $mode="Ent
 	SOY2::import("site_include.plugin.util_multi_language.domain.MultiLanguage".$mode."RelationDAO");
 	$dao = SOY2DAOFactory::create("MultiLanguage".$mode."RelationDAO");
 
-	foreach($_POST["multi_language"] as $lang => $_id){
-		$idx = SOYCMSUtilMultiLanguageUtil::getLanguageIndex($lang);
-		$_id = (int)$_id;
-		// 登録
-		if($_id > 0){
-			$obj = ($mode == "Entry") ? new MultiLanguageEntryRelation() : new MultiLanguageLabelRelation();
-			$obj->setParentId($id);
-			$obj->setLang($idx);
-			$obj->setChildId($_id);
-			
-			try{
-				$dao->insert($obj);
-			}catch(Exception $e){
+	if(isset($_POST["multi_language"]) && is_array($_POST["multi_language"]) && count($_POST["multi_language"])){
+		foreach($_POST["multi_language"] as $lang => $_id){
+			$idx = SOYCMSUtilMultiLanguageUtil::getLanguageIndex($lang);
+			$_id = (int)$_id;
+			// 登録
+			if($_id > 0){
+				$obj = ($mode == "Entry") ? new MultiLanguageEntryRelation() : new MultiLanguageLabelRelation();
+				$obj->setParentId($id);
+				$obj->setLang($idx);
+				$obj->setChildId($_id);
+				
+				try{
+					$dao->insert($obj);
+				}catch(Exception $e){
+					try{
+						$dao->delete($id, $idx);
+						$dao->insert($obj);
+					}catch(Exception $e){
+						//
+					}
+				}
+
+			// 削除
+			}else{
 				try{
 					$dao->delete($id, $idx);
-					$dao->insert($obj);
 				}catch(Exception $e){
 					//
 				}
 			}
-
-		// 削除
-		}else{
-			try{
-				$dao->delete($id, $idx);
-			}catch(Exception $e){
-				//
-			}
 		}
 	}
-
+	
 	return true;
 }
 
@@ -41,11 +43,13 @@ function multi_language_execute_common_remove_process(array $ids, string $mode="
 	SOY2::import("site_include.plugin.util_multi_language.domain.MultiLanguage".$mode."RelationDAO");
 	$dao = SOY2DAOFactory::create("MultiLanguage".$mode."RelationDAO");
 
-	foreach($ids as $id){
-		try{
-			$dao->deleteByParentId($id);
-		}catch(Exception $e){
-			//
+	if(count($ids)){
+		foreach($ids as $id){
+			try{
+				$dao->deleteByParentId($id);
+			}catch(Exception $e){
+				//
+			}
 		}
 	}
